feat: remove old gemini service and include ollama service
